requirements.txt
redis
requests
aioredis>=1.3.1,<2.0.0
# aioredis==2.0.1
aiohttp 
bitstring
# aiofcm
git+https://github.com/perishllc/aiofcm@main
python-dotenv
pylint
uvloop
python-socketio==4.6.1
websockets
python-rapidjson
aiohttp_cors
ed25519-blake2b
nanolib
nanopy